rgw: Prefix "nss db path" with "rgw"
If someone attempts to follow the existing documentation to
modify vstart's ceph.conf to operate with Keystone, the result
looks like this:

./src/common/ceph_crypto.cc: 73: FAILED assert(crypto_context != __null)

This occurs because "nss" is a valid family of prefixes, which
affects processes other than RADOS Gageway; in addition, the
configuration has to be written in the [global] section, since
vstart does not supply the -n argument.

The proposed fix is to add "nss" in the documentation.

Signed-off-by: Pete Zaitcev <zaitcev@redhat.com>