When you run a container with ceph it will fail with the error: ``` failed: '/bin/systemd-run -r bash -c 'ulimit -n 32768; /usr/bin/ceph-osd -i 2 --pid-file /var/run/ceph/osd.2.pid -c /etc/ceph/ceph.conf --cluster ceph -f'' Failed to create bus connection: No such file or directory ```
It's unlikely that people are running systemd within containers and thus i added a check if the processes is running in a container when systemd-run is triggered, if docker is detected the usual startup process is executed.

I chose to use grep -q to detect docker because it works on every system (tested it on privileged containers).

Signed-off-by: Rogier Dikkes <r.dikkes@gmail.com>