global/global_init: chown log, asok if drop privs is deferred
If we are deferring the drop privileges, then we are still root
and need to explicitly chown the log file and admin socket.

Note that this is a fragile solution: if there are other files
that we create or open for write between now and when privs are
eventually dropped, we need to explicitly handle them, too.

Signed-off-by: Sage Weil <sage@redhat.com>