ceph.spec.in: specify local sources
Tarball must be specified as a source in the spec file, and can either be
specified as an URL (for automated download at build time) or as a local file.
Since we are generating our own tarball from git at build time, we need a local
file.

Also, the Factory checker demands that ceph-rpmlintrc be listed as a
source file.

Signed-off-by: Nathan Cutler <ncutler@suse.com>