``` failed: '/bin/systemd-run -r bash -c 'ulimit -n 32768; /usr/bin/ceph-osd -i 2 --pid-file /var/run/ceph/osd.2.pid -c /etc/ceph/ceph.conf --cluster ceph -f'' Failed to create bus connection: No such file or directory ```
It's unlikely that people are running systemd within containers and thus i added a check if the processes is running in a container when systemd-run is triggered, if docker is detected the usual startup process is executed.

I messed up PR 7566, sorry for that!

Now checking for /.dockerinit if it exists, if the file does not exist we perform systemd-run

Signed-off-by: Rogier Dikkes <rogier.dikkes@surfsara.nl>