Makefile: workaround an automake bug for "make check"
override the automake variable of am__is_gnu_make. it is a workaround
for http://debbugs.gnu.org/cgi/bugreport.cgi?bug=18744. this bug was
fixed in automake 1.15, but automake 1.13 is supported. so we can not
just require 1.15 using `AM_INIT_AUTOMAKE`. this only happens when we
have lots of source files *and* we are using automake v1.14 or lower.

Fixes: #14723
Signed-off-by: Kefu Chai <kchai@redhat.com>