src/test/erasure-code/Makefile.am: avoid versioning shared libraries
These shared libraries are included in the ceph RPM package, which is not a
devel package and therefore should not include versioned shared libraries.

Also, the versions produced for these libraries do not seem to be of any
use. For example:

libec_example.so
libec_example.so.0
libec_example.so.0.0.0

http://tracker.ceph.com/issues/14756 Fixes: #14756

Signed-off-by: Nathan Cutler <ncutler@suse.com>