mds: introduce fs_cluster_id_t type
Unique identifier for a filesystem within
the ceph cluster.  Abbreviated to `fscid`,
deliberately chosen to not conflict
with legacy "fsid" identifiers.

Additionally add mds_role_t type which is
a more general form of rank: ranks are
now only unique within a particular fscid.
Roles are a fscid plus a rank.

Signed-off-by: John Spray <john.spray@redhat.com>