os/filestore: fix wrong scope of result code for error cases during mkfs
During the mkfs process, if we fail to read and sometimes rewrite the
op_seq, the mkfs process is abnormally terminated but we may
still return a positive answer to the caller, which is misleading.

This patch is provided to solve the above problems by correctly setting
'ret' correspondingly under these cases, which is the final answer to
be passed out for caller.

Fixes: #14814
Signed-off-by: xie xingguo <xie.xingguo@zte.com.cn>