os/filestore: fix wrong scope of result code for error cases during mount
This pr is similar to https://github.com/ceph/ceph/pull/7704
but I feel like making a new one in case they need to be treated separately,
such as one needs to be backported while the other does not, and vice versa.

During the mount process, if we fail to read op_seq or create a no-snap-fn,
the mount process is abnormally terminated but we may
still return a positive answer to the caller, which is misleading.

This patch is provided to solve the above problems by correctly setting
'ret' correspondingly under these cases, which is the final answer to
be passed out for caller.

Fixes: #14815
Signed-off-by: xie xingguo <xie.xingguo@zte.com.cn>