os/filestore: fix result code overwritten for clone
During the clone process, the result code of syscall-fstat which is intend for
getting the source size for copying is not checked and will be overwritten
by the following call, which is of potential risk and thus should be considered
as unsafe.

This pr solves the above problem by adding result checking against failure.
Also result code is reset to errno for better tracing if the ftruncate syscall
failed.

Fixes: #14817
Signed-off-by: xie xingguo <xie.xingguo@zte.com.cn>