pybind - PEP 440 fix, also out of tree build fix.
Instead of Looking for ceph_ver.h, this looks for .git_version .
Same information basically.  Ideally this should know how to look in
the build tree for ceph_ver.h, but that was more python mongery than I
wanted to hack.

Once this gets the result of "git describe", I've added logic to try
to format the version string to be PEP 440 friendly.  So, instead of
10.0.3-1027-ge51fd74, this will make 10.0.3.post10+ge51fd74 .  This fixes
a warning message at build time for me.

Signed-off-by: Marcus Watts <mwatts@redhat.com>