OSD: put a guard for updating heartbeat peers
As get_next_up_osd_after() and get_previous_up_osd_before() may refer to the same
osd, so we need to put a guard in case we try to add the same osd into want.
Note that the STL set can handle the above case, so this commit should be only
considered as cleanup, not a big deal.

Signed-off-by: xie xingguo <xie.xingguo@zte.com.cn>