mds: set multiversion inode's first
For multiversion inode, inode's 'first' can be different to its
parent dentry's 'first'. But when encoding a primary dentry, only
dentry's first gets stored. So when decoding a primary dentry,
we need to reconstruct inode's 'first'.

This patch uses old_inodes to calculate inode's first. It's not
perfect, but should be good enough.

Signed-off-by: Yan, Zheng <zyan@redhat.com>