New rgw rest api for bucket ops log
We can only use the command "radosgw-admin op list" to show bucket ops
log. After this we can use this api to download the log. The api is for users that
have no privilege to run radosgw-admin.

This api syntax is as followed:
GET
/?oplog&format=json&start-time=2016-03-01+02:00:00&end-time=2016-03-01+03:00:00
HTTP/1.1
Host: bucketname.rgw.com
Date: date
Authorization: authorization string

Signed-off-by: sunspot <sunspot0105@gmail.com>