civetweb / assert.h fix: XXX is this still necessary?
This was: automake changes for new civetweb.
but the automake part went away.

Adam C. Emerson <aemerson@redhat.com> says this about changes to
src/include/assert.h :

Admittedly kind of a gross way of doing it, but it's not the end of the world.
Doing it more properly would require taking the build system apart and putting
it back together which I don't think we want to do at this point in time.

Signed-off-by: Marcus Watts <mwatts@redhat.com>