osd: make os_flags an option
In one of our test environments an osd is unable to back to work
due to the journal is totally unrecoverable. The os_flags field
is introduced to handle such a case but never be made an option
and visible to normal user.

This pr tries to make os_flags field a configurable option and
no flags is enabled by default and thus shall cause no compatibility
relevant issues.

Signed-off-by: xie xingguo <xie.xingguo@zte.com.cn>