mon: OSDMonitor: wait proposal finished if thrash succeeds
By returning true, the thrash() method indicates caller that
it has modified the pending_inc and we shall waiting for the
changes to take effect, which is the most normal behaviour when we
receive and successfully perform a command from caller.

The problem here is that we discard the result of thrash() method
unconditionly here, and thus we shall always fall through to "reply"
instead of "update", which is the expecting action when thrash()
succeeds as mentioned above.

Signed-off-by: xie xingguo <xie.xingguo@zte.com.cn>