crush: remove straw_calc_version check from has_*_tunables()
Since we do not set this when we set tunables, we should not check it
when we report which tunables they have set.  For example, a savvy
operator on firefly tunables should set straw_calc_version to 1 but we
still want the cluster to report that they are on firefly tunables.

Signed-off-by: Sage Weil <sage@redhat.com>