debian: move ceph_daemon.py to ceph-common where it belongs
The ceph_daemon.py file was introduced in infernalis by c3ef6409 -
unnoticed for some time was the fact that this original commit put the file
into two different packages: the Debian packaging had it in "ceph", while
the RPM packaging had it in "ceph-common".

As a result, Debian and Ubuntu users who installed only ceph-common were
unable to use the "daemon" and "daemonperf" subcommands of the ceph
executable.

This was fixed by 0cbe3dea which was merged in time for jewel. This commit
can be considered a backport of 0cbe3dea in the sense that it fixes the
same issue in infernalis. However, 0cbe3dea is unsuitable for
cherry-picking because it splits the ceph package into four smaller
packages (the so-called "package split" feature which is not being
backported to infernalis).

http://tracker.ceph.com/issues/15048 Fixes: #15048

Signed-off-by: Nathan Cutler <ncutler@suse.com>