LFNIndex::lfn_translate: consider alt attr as well
If the file has an alt attr, there are two possible matching
ghobjects.  We want to make sure we choose the right one for
the short name we have.  If we don't, a split while there are
two objects linking to the same inode will result in one of
the links being orphaned in the source directory, resulting
in #14766.

Signed-off-by: Samuel Just <sjust@redhat.com>
(cherry picked from commit 850ce11ea3c4dacc6ff1a3ce7ff37000818267c2)