rgw: Remote the new rgw_swift_account_in_url option
A recent commit introduced rgw_swift_account_in_url, which appears
to be not necessary. This patch uses token format to pass the
information that current code extracts from the option. This way
a smooth upgrades without undue configuration are possible, and
the administrative load is not raised. No need to document what
software can do automatically either.

N.B. We need to find out of TempUrl continues to function without
this option. Textually it does.

Signed-off-by: Pete Zaitcev <zaitcev@redhat.com>