mon/PGMonitor: reliably mark PGs state
In beb99d708a5051405a32f46fc10cef473bfddce3 we partially addressed
the case where a acting_primary update raced with an attemp to
mark the pg state.  However, if the acting_primary changed, we
didn't check the new acting_primary to ensure that it is also
down and we should still mark things stale.

Also, printing out the acting_primary aids debugging.

Signed-off-by: Sage Weil <sage@redhat.com>