mon: ignore msg without session attached
there is chance that a connection gets reset, and the session attached
to a message that was sent over this connection is reset. i.e.
con->set_priv(NULL) in Monitor::ms_handle_reset(), but we are about to
handle_command() this message after the message is somehow requeued. so,
not all MonOpRequests have an attached Session, especially those are
re-dispatched. we could check conn->is_connected() everywhere, but
that's error-prone and hard to maintain. so, in this change:

* we need to check for the session before looking at it closer.
* hold a reference to it when holding it, because in another thread,
  ms_handle_reset() could be freeing it.

Fixes: #15113
Signed-off-by: Kefu Chai <kchai@redhat.com>