vstart: support creating multiple cephfs filesystems
This is handy if you want to quickly create a cluster
with lots of filesystems, for playing with e.g. what
the ceph status output looks like when multiple filesystems
are in play.

This changes the default filesystem/pool names to be
e.g. cephfs_data_a instead of just `cephfs_data`, but anything
who relied on those names in a vstart environment houldn't have
been (this won't affect teuthology tests, of course, where we
still have a few annoying hardcoded pool names).

Signed-off-by: John Spray <john.spray@redhat.com>