mon/MonClient: make authenticate timeout more reliable
The LibRadosMiscConnectFailure.ConnectFailure test occasionally
fails with a .000001 timeout.  I'm pretty sure it's because this
cond always manages a signal queued by the time it goes to wait,
or gets a signal before the deadline.

Fix by adding an outter check as well so that we'll error out
regardless of the queued signal vs timeout check order in the
cond.

Fixes: #13392
Signed-off-by: Sage Weil <sage@redhat.com>