Adding post-install (ceph-mon.postinst and ceph-osd.postinst) and pre-removal (ceph-mon.prerm and ceph-osd.prerm) scripts for ceph-mon and ceph-osd Debian packages.
These scripts are derived from correponding ceph-mds post-install/pre-removal
scripts, and are triggered right after the package installation/removal.

The post-install scripts are required to mark the ceph-mon-all and
ceph-osd-all upstart jobs in start/running state after installing packages.

Similarly, the pre-removal scripts are required to mark the ceph-mon-all and
ceph-osd-all upstart jobs in stop/waiting state after removing packages.

We need to have the ceph-mon-all and ceph-osd-all jobs in start/running state
for a stop event send to ceph-all stop the running monitors and osds.
Currently, even though ceph-all and ceph-mon are in start/running state after
a 'ceph-deploy mon create' operation (which sends an emit event to ceph-mon),
a subsequent 'stop ceph-all' operation fails to stop the monitors. Similarly,
for ceph-osd as well.

Signed-off-by: Vishal Agarwal <vishal.agarwal@sandisk.com>