mon/PGMonitor: check_down_pgs against the consumed OSDMap
This ensures that we are in sync with pg_map.pg_by_osd etc.

Strengthen/fix the check in the second loop now that we are in sync:
if we found it in the pg_by_osd map, then we can assert that
the acting_primary matches, or else there is a bug.

Signed-off-by: Sage Weil <sage@redhat.com>