mds: simplify standby/standby-replay logic
This used to use an arcane set of constants
in standby_for_rank, combined with daemons sometimes
sending requests to enter state STANDBY_REPLAY.

Simplify this so that there is only one place we
put daemons into standby replay, and that is
in tick->maybe_promote_staandby.

There is a behavioural change in here, which is
that we used to sometimes promote standbys to
be standby-replay, even if they didn't have
"standby replay = true", when they did have
a standby_for_rank or standby_for_name set.
I'm calling that a bug, and making it so that
daemons will only go into standby-replay if
"standby replay = true" is set.

Signed-off-by: John Spray <john.spray@redhat.com>
(cherry picked from commit 2a2a56b2d0e9e4224cae27f80d990b787e209bed)