tests: Removing one ceph-dencoder call in check-generated.sh
The first ceph-dencoder call is very unlikely to fail and represent a
bottleneck as the parallel computation are only done one this test is
completed.

The idea of this patch is to run immediately the 4 dencoders process in
parallel and check the resulting error code. If one fail then we report the
failure.

As the failure is very unlikely, that saves time and makes the code
simplier too.

Signed-off-by: Erwan Velu <erwan@redhat.com>