test/throttle: add usleep for Thread_get::entry
In the test case ThrottleTest.get, we try to delay one or a few microseconds
to do throttle.put that will wake up a waited throttle,
and we expect t.waited is false for short delay and true for long delay,
however the result did not meet our expectations
since even one microseconds is too long.
Maybe we should add usleep for Thread_get::entry.

Signed-off-by: Yan Jun <yan.jun8@zte.com.cn>