osd: force update pg_stats for duplicated publish_stats_to_osd()
We may call publish_stats_to_osd() multiple times during boot/peering/recovering,
in which cases we may want update some important info such as pg state and report
it to user on time. However the statistics may stay in the pg_stat_queue for a
long time in some abnormal cases, and we shall send the latest statistics instead,
which means a force update of existing pg_stats in the pg_stat_queue is mandatory.

Also, simplify pg_stat_queue_lock a little, which is good for extensibility.

Signed-off-by: xie xingguo <xie.xingguo@zte.com.cn>