Makefile-env.am: set a default for CEPH_BUILD_VIRTUALENV (part 3)
The default values in the run-tox.sh scripts must be updated as well, so
that it is possible to run them from the command line for debug
purposes. Otherwise it is mandatory to run them from

  make TESTS=ceph-detect-init/run-tox.sh check

which introduces a significant delay. The alternative is to run with

  CEPH_DETECT_INIT_VIRTUALENV:=/tmp/ceph-detect-init-virtualenv \
    ceph-detect-init/run-tox.sh

but it's difficult to figure out.

Followup of fa450896ace454dbd140b2d4c71535ab015fc407

Signed-off-by: Loic Dachary <loic@dachary.org>