mds: fix message leak during handle_core_message()
On successfully handing a core message,  the handle_core_message()
method shall return true and is responsible for dropping the corresponding
message's reference to get it correctly released.

However, if we receive a CEPH_MSG_OSD_MAP message, we won't drop
its reference unless it is neither from a monitor nor an OSD, which means
in most cases we won't release this kind of message correctly.

This pr solves the above problem by dropping the message reference
appropriately.

Signed-off-by: xie xingguo <xie.xingguo@zte.com.cn>