mds: drop message reference on error exit
From the call stack handle_core_message()->handle_mds_map(),
handle_core_message() expects that handle_mds_map() will definitely
take good care of the CEPH_MSG_MDS_MAP message and will drop
the message reference on returning.

However, in the belowing code path, it won't do such a tidy up thing.
Although this one is not necessary because we are going to shutdown
anyway, but it is a good habit to do the cleanup.

Signed-off-by: xie xingguo <xie.xingguo@zte.com.cn>