build: change ccache invocation to path style
If ccache is invoked as two arguments eg CXX='ccache g++', then pybind
linking can fail because of long-standing distutils bug 8027. Some
distributions patch distutils for this, but it's not consistent.

The alternative way to use ccache is by altering PATH, but it needs
slight care because the actual directory for the ccache symlinks is not
consistent between distributions.

X-Distutils-Bug-URL: https://bugs.python.org/issue8027
Signed-off-by: Robin H. Johnson <robin.johnson@dreamhost.com>