osd/ReplicatedPG: fix potential memory leak
During do_pg_op() process, theoretically we are possible to bind a
PGLS/PGNLS_FILTER command and a PG_HITSET_GET command. We may successfully
handled the former while failing to handle the latter due to the HitSet
object is currently unreadable. We will delay and replay this op later
as a result, which therefore shall cause memory leak here, as "filter"
is not sucessfully relased under this case.

The above case maybe is rare but this change is defensive and makes code
strong and thus shall be considered as worthwhile.

Signed-off-by: xie xingguo <xie.xingguo@zte.com.cn>