Drop a redundant warning about experimental features
The problem this patch addresses is the insane amount of
warning messages produced by our test scripts, such as
vstart.sh.

One might think that this warning delivers value to a user
who runs a script that, unbeknownst to the user, enables
experimental features. This, as it turns out, is redundant.
We do not only print the warning that this patch removes,
but also a pointed warning where the feature flag is used
(see CephContext::check_experimental_feature_enabled).
So, all this time, we endured the flood of warnings for
nothing.

Signed-off-by: Pete Zaitcev <zaitcev@redhat.com>