rgw_admin: improve period update errors
If we do a period update on a realm with multiple masters, we fail
stating invalid argument, improving this to suggest that we have
actually errored out because of multiple master zg in a realm, and
printing both of these zonegroups. RGWPeriodMap::update takes
CephContext as an argument, so that we can use ldout. Also dropping the
error messages reporting invalid arguments on period.update() as these
are handled within the function itself whenever it manifests itself.

Fixes: http://tracker.ceph.com/issues/15251
Signed-off-by: Abhishek Lekshmanan <abhishek@suse.com>