osd: fix watch reconnect race
It's possible to process a watch reconnect op, setting Watch::conn, before
the ms_handle_reset path tries to disconnect the old one.  Since reset()
was blindly disconnecting the "current" connection, we could race and
disconnect the new con instead of the old one.

Fix this by specifying which con to disconnect.

Fixes: http://tracker.ceph.com/issues/15441
Signed-off-by: Sage Weil <sage@redhat.com>