hammer: make dist: include rocksdb sources
make dist does not recurse into SUBDIRS if DIST_SUBDIRS are defined.
As a result rocksdb source was missing in the tarball if the source
has been configured --with-librocksdb-static. Unconditionally add
rocksdb to DIST_SUBDIRS to avoid the problem. As a side effect files
rocksdb's configure script, Makefile.in, and various autoconf/libtool
bits are also included in the dist tarball.

Signed-off-by: Alexey Sheplyakov <asheplyakov@mirantis.com>