rgw: Add a test for multi-tenancy
This is a very rudimentary test for now. For one thing, it
only creates a tenantized user, but does not even try to run
the radosgw and access it with Boto. And it already has several
XXX tags. But baby steps.

Two areas of exising code had to be fixed up.

 - We forgot to patch an argument through in RGWcluster.rgw_admin()

 - We must not process the arguments of a command through eval(1)
   in test-rgw-call.sh. This causes a double-eval with the function
   x() in test-rgw-common.sh, and as a result arguments with spaces
   or dollars cannot be used. There is no way to escape or quote.
   But we need the dollar for the tenant!

Signed-off-by: Pete Zaitcev <zaitcev@redhat.com>